+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.5] - 2021-11-22
+### Added
+- Unwrap link
+### Removed
+- New tab proxy option
+### Fixed
+- Missing secret key error
+
 ## [0.4.4] - 2021-11-15
 ### Changed
 - Prevent importing accounts without keys
@@ -100,8 +109,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Storage encryption
 - Docs
 
-[Unreleased]: https://github.com/adshares/ads-browser-wallet/compare/v0.4.4...HEAD
+[Unreleased]: https://github.com/adshares/ads-browser-wallet/compare/v0.4.5...HEAD
 
+[0.4.5]: https://github.com/adshares/ads-browser-wallet/compare/v0.4.4...v0.4.5
 [0.4.4]: https://github.com/adshares/ads-browser-wallet/compare/v0.4.3...v0.4.4
 [0.4.3]: https://github.com/adshares/ads-browser-wallet/compare/v0.4.2...v0.4.3
 [0.4.2]: https://github.com/adshares/ads-browser-wallet/compare/v0.4.1...v0.4.2